LFLT-139 Populate bridge with required endpoints - Added integration tests; small fixes.

+++ src/main/java/hu/psprog/leaflet/bridge/service/EntryBridgeService.java
@@ -48,7 +48,7 @@ public interface EntryBridgeService {
      * @return page of public entries under given category
      * @throws CommunicationFailureException if client fails to reach backend application
      */
-    WrapperBodyDataModel<EntryListDataModel> getPageOfPublicEntriesByCategory(int categoryID, int page, int limit, OrderBy.Entry orderBy, OrderDirection orderDirection)
+    WrapperBodyDataModel<EntryListDataModel> getPageOfPublicEntriesByCategory(Long categoryID, int page, int limit, OrderBy.Entry orderBy, OrderDirection orderDirection)
             throws CommunicationFailureException;
 
     /**
